テストコードを整理する
