+++ pkg/infrastracture/repository/summaries_test.go
@@ -2,7 +2,6 @@ package repository
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"sort"
 	"testing"
@@ -16,79 +15,6 @@ import (
 	"github.com/shoet/webpagesummary/pkg/infrastracture/entities"
 	"github.com/shoet/webpagesummary/pkg/testutil"
 )
-
-func TestMain(m *testing.M) {
-	_ = MustSetup()
-	m.Run()
-	MustTeardown(&TeardownInput{})
-}
-
-var tableName = "web_page_summary"
-
-type SetUpOutput struct{}
-
-type TeardownInput struct{}
-
-func MustSetup() *SetUpOutput {
-	ctx := context.Background()
-	awsConfig, err := testutil.NewAwsConfigWithLocalStack(ctx)
-	if err != nil {
-		panic(fmt.Sprintf("failed load aws config: %s\n", err.Error()))
-	}
-	if err := testutil.CreateDynamoDBForTest(ctx, *awsConfig, &dynamodb.CreateTableInput{
-		TableName: &tableName,
-		AttributeDefinitions: []types.AttributeDefinition{
-			{
-				AttributeName: aws.String("id"),
-				AttributeType: types.ScalarAttributeTypeS,
-			},
-			{
-				AttributeName: aws.String("task_status"),
-				AttributeType: types.ScalarAttributeTypeS,
-			},
-		},
-		GlobalSecondaryIndexes: []types.GlobalSecondaryIndex{
-			{
-				IndexName: aws.String("StatusIndex"),
-				KeySchema: []types.KeySchemaElement{
-					{
-						AttributeName: aws.String("task_status"),
-						KeyType:       types.KeyTypeHash,
-					},
-				},
-				Projection: &types.Projection{
-					ProjectionType: types.ProjectionTypeAll,
-				},
-			},
-		},
-		KeySchema: []types.KeySchemaElement{
-			{
-				AttributeName: aws.String("id"),
-				KeyType:       types.KeyTypeHash,
-			},
-		},
-		BillingMode: types.BillingModePayPerRequest,
-	}); err != nil {
-		var e *types.ResourceInUseException
-		if errors.As(err, &e) {
-			fmt.Printf("table already exists: %s\n", tableName)
-			return nil
-		}
-		panic(fmt.Sprintf("failed create dynamodb: %s\n", err.Error()))
-	}
-	return &SetUpOutput{}
-}
-
-func MustTeardown(input *TeardownInput) {
-	ctx := context.Background()
-	awsConfig, err := testutil.NewAwsConfigWithLocalStack(ctx)
-	if err != nil {
-		panic(fmt.Sprintf("failed load aws config: %s\n", err.Error()))
-	}
-	if err := testutil.DropDynamoDBForTest(ctx, *awsConfig, tableName); err != nil {
-		panic(fmt.Sprintf("failed delete sqs queue: %s\n", err.Error()))
-	}
-}
 
 func Test_SummaryRepository_GetSummary(t *testing.T) {
 	ctx := context.Background()
